Add navigation to Dashboard from UserDropdown
A new DropdownMenuItem was added to the UserDropdown function in chat.tsx to allow users to navigate directly to the Dashboard. This is achieved by constructing a URL using the NEXT_PUBLIC_WEBAPP_URL environment variable, falling back to "https://app.localtest.local:3000" if the variable is not set. The user is then redirected to this URL when the new menu item is selected.